chore: update dependencies and refactor project structure
- Updated the  to downgrade  and  packages.
- Commented out the  file in  to prevent loading issues.
- Refactored import paths in various files to align with the new project structure.
- Removed unused files and widgets to clean up the codebase, including several widget files and use case implementations.
- Added TODO comments for future implementation of environment configuration and authentication logic.
