+++ lib/core/presentation/widgets/category_selection_widget.dart
