+++ lib/core/presentation/widgets/inter_text.dart
