+++ lib/core/presentation/widgets/result_count_widget.dart
