+++ lib/core/presentation/widgets/search_bar_widget.dart
