+++ lib/features/auth/presentation/cubit/auth_cubit.dart
@@ -1,12 +1,10 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nashik/core/auth/google_auth_service.dart';
-import 'package:nashik/core/router/app_router.dart';
 import 'package:nashik/core/supabase/config.dart';
-import 'package:nashik/features/auth/domain/usecases/get_current_user.dart';
-import 'package:nashik/features/auth/domain/usecases/signin_with_email.dart';
-import 'package:nashik/features/auth/domain/usecases/signup_with_email.dart';
+import 'package:nashik/features/auth/domain/use_cases/get_current_user.dart';
+import 'package:nashik/features/auth/domain/use_cases/signin_with_email.dart';
+import 'package:nashik/features/auth/domain/use_cases/signup_with_email.dart';
 import 'package:nashik/features/auth/presentation/cubit/auth_state.dart';
-import 'package:nashik/features/auth/presentation/pages/oauth_callback_page.dart';
 import 'package:supabase_flutter/supabase_flutter.dart' hide AuthState;
 
 class AuthCubit extends Cubit<AuthState> {
@@ -23,30 +21,35 @@ class AuthCubit extends Cubit<AuthState> {
   }
 
   void _initializeAuthState() {
-    SupabaseConfig.client.auth.onAuthStateChange.listen((data) {
-      final AuthChangeEvent event = data.event;
-      final Session? session = data.session;
+    // TODO: Uncomment when Supabase is initialized
+    // SupabaseConfig.client.auth.onAuthStateChange.listen((data) {
+    //   final AuthChangeEvent event = data.event;
+    //   final Session? session = data.session;
 
-      if (event == AuthChangeEvent.signedIn && session != null) {
-        try {
-          final currentLocation = Approuter.router.location;
-          if (!currentLocation.contains(OAuthCallbackPage.routePath)) {
-            _loadCurrentUser();
-          }
-        } catch (e) {
-        _loadCurrentUser();
-        }
-      } else if (event == AuthChangeEvent.signedOut) {
-        emit(const AuthState.unauthenticated());
-      }
-    });
+    //   if (event == AuthChangeEvent.signedIn && session != null) {
+    //     try {
+    //       final currentLocation = Approuter.router.location;
+    //       if (!currentLocation.contains(OAuthCallbackPage.routePath)) {
+    //         _loadCurrentUser();
+    //       }
+    //     } catch (e) {
+    //     _loadCurrentUser();
+    //     }
+    //   } else if (event == AuthChangeEvent.signedOut) {
+    //     emit(const AuthState.unauthenticated());
+    //   }
+    // });
 
-    final user = SupabaseConfig.client.auth.currentUser;
-    if (user != null) {
-      _loadCurrentUser();
-    } else {
-      emit(const AuthState.unauthenticated());
-    }
+    // TODO: Uncomment when Supabase is initialized
+    // final user = SupabaseConfig.client.auth.currentUser;
+    // if (user != null) {
+    //   _loadCurrentUser();
+    // } else {
+    //   emit(const AuthState.unauthenticated());
+    // }
+    
+    // Temporary: Set unauthenticated state when Supabase is not initialized
+    emit(const AuthState.unauthenticated());
   }
 
   Future<void> _loadCurrentUser() async {
