+++ lib/main.dart
@@ -5,12 +5,8 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:nashik/android_app.dart';
-import 'package:nashik/core/deep_link/deep_link_service.dart';
 import 'package:nashik/core/di/get_it.dart';
-import 'package:nashik/core/env/config.dart';
 import 'package:nashik/core/router/app_router.dart';
-import 'package:nashik/core/supabase/config.dart';
-import 'package:nashik/features/auth/presentation/cubit/auth_cubit.dart';
 import 'package:nashik/features/profile/presentation/cubit/profile_cubit.dart';
 import 'package:nashik/ios_app.dart';
 
@@ -19,20 +15,24 @@ import 'package:nashik/ios_app.dart';
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
+  // TODO: Uncomment when .env file is added to assets
   // Initialize Config (loads .env file)
-  await Config.instance;
+  // await Config.instance;
 
+  // TODO: Uncomment when .env file is added to assets
   // Initialize Supabase with environment variables
-  await SupabaseConfig.initialize();
+  // await SupabaseConfig.initialize();
 
   // Initialize service locator (encrypted secure storage)
   await serviceLocatorInit();
 
+  // TODO: Uncomment when Supabase is initialized
   // Initialize deep linking
-  DeepLinkService().initialize();
+  // DeepLinkService().initialize();
 
+  // TODO: Uncomment when Supabase is initialized
   // Process any initial deep link (e.g., app opened via deep link)
-  await DeepLinkService().getInitialLink();
+  // await DeepLinkService().getInitialLink();
 
   // initialize the go router
   Approuter.init();
@@ -49,7 +49,8 @@ class MyApp extends StatelessWidget {
       designSize: const Size(360, 920),
       child: MultiBlocProvider(
         providers: [
-          BlocProvider(create: (context) => locator<AuthCubit>()),
+          // TODO: Uncomment when Supabase is initialized
+          // BlocProvider(create: (context) => locator<AuthCubit>()),
           BlocProvider(create: (context) => ProfileCubit()),
         ],
         child: Platform.isIOS ? const IosApp() : const AndroidApp(),
